refactor: js 디렉토리 구조 개선
components 폴더 아래 header, inputBar 폴더를 추가함.
js 파일이 많아지면 요소별로 구별이 안 되기 때문임.
폴더 이름은 파일과 일관성을 유지하기 위해 camelCase로 지음.
