+++ README.md
@@ -58,7 +58,7 @@
 - 내용
 - 결제수단
     - [ ] 클릭하면 드롭다운 패널 등장
-    - [ ] 드롭다운 패널을 벗어나면 패널을 안 보이게 off
+    - [ ] 마우스가 드롭다운 패널을 벗어나면 닫기
     - [ ] x버튼 클릭하면 알림창 띄우기
         - [ ] 클릭시 아이콘을 줄였다 늘려 클릭되는 느낌 주기
         - [ ] 배경 딤 처리
