+++ index.html
@@ -27,9 +27,9 @@
         <section class="month-carousel">
           <button type="button" class="button--transparent prev-month-btn"></button>
           <section class="month-view">
-            <p>2023</p>
-            <p>2</p>
-            <p>February</p>
+            <p></p>
+            <p></p>
+            <p></p>
           </section>
           <button type="button" class="button--transparent next-month-btn"></button>
         </section>
