+++ src/js/components/header/monthCarousel.js
