+++ src/js/components/inputBar/priceInput.js
